OCC-71: Restructure application
